% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs_form_csv.R
\name{costs_form_csv}
\alias{costs_form_csv}
\title{Download Costs form entries using EpiCollect5 API}
\usage{
costs_form_csv()
}
\value{
The dataset is written as a csv file in the IKEA_Ethopian_grain_mill_data folder on local storage
}
\description{
Download Costs form entries using EpiCollect5 API
}
\examples{
#Download and store the Revenue and Costs form of the IKEA Ethiopia grain mill project in CSV
\dontrun{
Costs_form <- costs_form_csv()
}
}
