% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_sensor_csv_files.R
\name{dates_n_labels}
\alias{dates_n_labels}
\title{Rename_dates_kwh}
\usage{
dates_n_labels(df)
}
\arguments{
\item{df}{one of the sensor files above eg. X085c}
}
\value{
add columns year_created
}
\description{
A function to for initial data cleaning of sensor files
}
\examples{
dates_n_labels(XO869)
}
\author{
Ama Owusu-Darko
}
