% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_responses.R
\name{get_survey_responses}
\alias{get_survey_responses}
\title{Download Revenue form using EpiCollect5 API}
\usage{
get_survey_responses()
}
\value{
The dataset is written as a csv file in the IKEA_Ethopian_grain_mill_data folder on local storage
}
\description{
Download Revenue form using EpiCollect5 API
}
\examples{
#Download and store the Revenue form of the IKEA Ethiopia grain mill project in CSV
\dontrun{
get_survey_responses()
}
}
