% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grain_operator_form_csv.R
\name{grain_operator_form_csv}
\alias{grain_operator_form_csv}
\title{Download Grain operator form entries using EpiCollect5 API}
\usage{
grain_operator_form_csv()
}
\value{
The dataset is written as a csv file in the IKEA_Ethopian_grain_mill_data folder on local storage
}
\description{
Download Grain operator form entries using EpiCollect5 API
}
\examples{
#Download and store the Grain operator form of the IKEA Ethiopia grain mill project in CSV
\dontrun{
Grain_operator_form<- grain_operator_form()
}
}
