% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenue_and_costs_form_csv.R
\name{revenue_costs_form_csv}
\alias{revenue_costs_form_csv}
\title{Download Revenue and Costs form entries using EpiCollect5 API}
\usage{
revenue_costs_form_csv()
}
\value{
The dataset is written as a csv file in the IKEA_Ethopian_grain_mill_data folder on local storage
}
\description{
Download Revenue and Costs form entries using EpiCollect5 API
}
\examples{
#Download and store the Revenue and Costs form of the IKEA Ethiopia grain mill project in CSV
\dontrun{
Revenue_costs_form <- revenue_costs_form_csv()
}
}
